moved panel actions
