+++ app/components/templates/panel/dashboard-entity-panel.tsx
@@ -8,15 +8,10 @@ import { type IDashboardPanelCreateEntityStrategy } from '#app/strategies/compon
 import { type IDashboardPanelDeleteEntityStrategy } from '#app/strategies/component/dashboard-panel/delete-entity.strategy'
 import { type IDashboardPanelUpdateEntityOrderStrategy } from '#app/strategies/component/dashboard-panel/update-entity-move.strategy'
 import { type IDashboardPanelUpdateEntityVisibleStrategy } from '#app/strategies/component/dashboard-panel/update-entity-visible.strategy'
-import {
-	SidebarPanel,
-	SidebarPanelRow,
-	SidebarPanelRowContainer,
-	SidebarPanelRowReorderContainer,
-} from '..'
-import { FormFetcherMoveIcon } from '../form/fetcher/move-icon'
+import { SidebarPanel, SidebarPanelRow, SidebarPanelRowContainer } from '..'
 import { PanelEntityRowActions } from './dashboard-entity-panel.actions'
 import { PanelEntityHeader } from './dashboard-entity-panel.header'
+import { PanelEntityRowReorder } from './dashboard-entity-panel.reorder'
 
 type PanelEntities = IDesignWithType[]
 type PanelEntity = IDesignWithType
@@ -118,49 +113,5 @@ export const PanelEntityRow = ({
 				/>
 			</SidebarPanelRowContainer>
 		</SidebarPanelRow>
-	)
-}
-
-export const PanelEntityRowReorder = ({
-	entity,
-	parentTypeId,
-	parent,
-	entityCount,
-	entityIndex,
-	strategyReorder,
-}: {
-	entity: PanelEntity
-	parentTypeId: designParentTypeIdEnum
-	parent: PanelEntityParent
-	entityCount: number
-	entityIndex: number
-	strategyReorder: IDashboardPanelUpdateEntityOrderStrategy
-}) => {
-	const atTop = entityIndex === 0
-	const atBottom = entityIndex === entityCount - 1
-
-	return (
-		<SidebarPanelRowReorderContainer>
-			<FormFetcherMoveIcon
-				entityId={entity.id}
-				parentTypeId={parentTypeId}
-				parentId={parent.id}
-				route={strategyReorder.route}
-				formId={strategyReorder.formId}
-				schema={strategyReorder.schema}
-				direction="up"
-				atTopOrBottom={atTop}
-			/>
-			<FormFetcherMoveIcon
-				entityId={entity.id}
-				parentTypeId={parentTypeId}
-				parentId={parent.id}
-				route={strategyReorder.route}
-				formId={strategyReorder.formId}
-				schema={strategyReorder.schema}
-				direction="down"
-				atTopOrBottom={atBottom}
-			/>
-		</SidebarPanelRowReorderContainer>
 	)
 }
